- improved the server startup and some of its internal organization - eslinting the server code, too
