+++ package.json
@@ -26,15 +26,16 @@
     "copy-static-files": "run-s copy:img copy:vendor:*",
     "l10n": "pontoon-to-json",
     "less": "lessc less/index.less dist/styles.css --source-map=dist/styles.css.map --clean-css=\"--s0\"",
-    "lint:fix": "eslint --ext .js --ext .jsx . -- --fix",
+    "lint:fix": "eslint --ext .js --ext .jsx --fix .",
+    "lint:server": "eslint --ext .js app.js ./server",
     "postinstall": "npm run prep",
     "prep": "run-s cleanup copy-static-files less l10n webpack:static webpack",
-    "start": "run-p start:*",
+    "start": "run-s webpack:* && run-p start:*",
     "start:less": "chokidar-cmd -c 'npm run less' -t less",
     "start:app": "npm run webpack -- --watch",
     "start:static": "npm run webpack:static -- --watch",
     "start:server": "node app",
-    "test": "npm run prep",
+    "test": "run-s prep lint:server",
     "webpack": "webpack --config ./config/webpack/webpack.client.config.js --progress --colors",
     "webpack:static": "webpack --config ./config/webpack/webpack.server.config.js --progress --colors"
   },
