+++ server/404.js
@@ -1,0 +1,13 @@
+var notFoundHTML = [
+  '<!doctype html>',
+  '<html><head>',
+  '<meta charset="utf-8">',
+  '<title>404 - Page not found</title>',
+  '</head><body>',
+  '<p>404 - Page not found</p>',
+  '</body></html>'
+].join('');
+
+module.exports = function(req, res, next) {
+  res.status(404).send(notFoundHTML);
+};
